rename: 메인 폴더 이름 변경
