+++ src/components/Main/TapViewerArea.js
